push

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: aderison <aderison@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/08 08:44:49 by arnaud            #+#    #+#             */
-/*   Updated: 2024/04/05 20:37:28 by aderison         ###   ########.fr       */
+/*   Updated: 2024/04/10 14:07:12 by aderison         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t size)
 	size_t	j;
 	size_t	len_src;
 	size_t	i;
-	int		total_len;
+	size_t	total_len;
 	size_t	len_dst;
 
 	if (!dst || !src)
